// src/platform/plugins/shared/discover/public/application/main/components/field_stats_table/types.d.ts
export interface FieldStatisticsTableProps {
    /**
     * Determines which columns are displayed
     */
    columns: string[];
    /**
     * The used data view
     */
    dataView: DataView;
    /**
     * Saved search description
     */
    searchDescription?: string;
    /**
     * Saved search title
     */
    searchTitle?: string;
    /**
     * Optional saved search
     */
    savedSearch?: SavedSearch;
    /**
     * Optional query to update the table content
     */
    query?: Query | AggregateQuery;
    /**
     * Filters query to update the table content
     */
    filters?: Filter[];
    /**
     * Callback to add a filter to filter bar
     */
    onAddFilter?: (field: DataViewField | string, value: string, type: '+' | '-') => void;
    /**
     * Metric tracking function
     * @param metricType
     * @param eventName
     */
    trackUiMetric?: (metricType: UiCounterMetricType, eventName: string | string[]) => void;
    /**
     * Search session id to save to or restore from
     */
    searchSessionId?: string;
    /**
     * If table should query using ES|QL
     */
    isEsqlMode?: boolean;
    /**
     * Time range
     */
    timeRange?: TimeRange;
    /**
     * Force refresh the table
     */
    lastReloadRequestTime?: number;
    /**
     * Whether to hide the preview/mini distribution chart in the table
     */
    hideAggregatedPreview?: boolean;
    /**
     * The total number of hits from the last fetch
     */
    totalHits?: number;
    /**
     * Callback to handle state updates
     */
    updateState?: (changes: Partial<DataVisualizerTableState>) => void;
}